cleanup inv relaying and mempool sync slowness, enable walletbackup pytest

+++ src/net/netconstants.h
@@ -132,7 +132,7 @@ static const int64_t BLOCK_DOWNLOAD_TIMEOUT_BASE = 1000000;
 static const int64_t BLOCK_DOWNLOAD_TIMEOUT_PER_PEER = 500000;
 
 /** Number of blocks that can be requested at any given time from a single peer. */
-static const int MAX_BLOCKS_IN_TRANSIT_PER_PEER = 64;
+static const int16_t MAX_BLOCKS_IN_TRANSIT_PER_PEER = 64;
 
 // SHA256("netgroup")[0:8]
 static const uint64_t RANDOMIZER_ID_NETGROUP = 0x6c0edd8036ef4036ULL;
